--- User.php
@@ -1,19 +1,0 @@
-<?php
-    class User{
-        
-        private $db;
-        
-        private $id;
-        private $user;
-        private $mail;
-        private $role;
-        
-        function __construct($userId){
-                $
-        }
-        
-        function getId(){
-            
-        }
-    }
-?>
